+++ packages/components/organisms/f-checkout/src/components/ErrorDialog.vue
@@ -7,7 +7,7 @@
     >
         <h3
             data-test-id="checkout-issue-modal-title"
-            class="u-noSpacing"
+            :class="$style['c-checkout-errorTitle']"
         >
             {{ $t(`errorMessages.checkoutIssues.${errorCode}.title`, { serviceType: serviceTypeText }) }}
         </h3>
@@ -90,6 +90,10 @@ export default {
 </script>
 
 <style lang="scss" module>
+.c-checkout-errorTitle {
+    margin: 0 spacing(x3);
+}
+
 .c-checkout-redirectButton {
     margin: spacing(x4) 0 spacing(x0.5);
 }