+++ packages/services/f-http/src/tests/createClient.test.js
@@ -1,3 +1,6 @@
+import axios from 'axios';
+import MockAdapter from 'axios-mock-adapter';
+
 import createClient from '../createClient';
 
 describe('createClient', () => {
@@ -29,7 +32,8 @@ describe('createClient', () => {
                 timeout: 10000,
                 errorCallback: null,
                 contentType: 'application/json',
-                instanceName: 'Generic Front End'
+                instanceName: 'Generic Front End',
+                isDevelopment: false
             };
 
             // Act
@@ -47,13 +51,15 @@ describe('createClient', () => {
             const expectedErrorCallback = () => {};
             const expectedContentType = 'application/mpeg';
             const expectedInstanceName = 'Test Test Test';
+            const expectedIsDevelopment = true;
 
             const expectedResult = {
                 baseUrl: expectedBaseUrl,
                 timeout: expectedTimeout,
                 errorCallback: expectedErrorCallback,
                 contentType: expectedContentType,
-                instanceName: expectedInstanceName
+                instanceName: expectedInstanceName,
+                isDevelopment: expectedIsDevelopment
             };
 
             // Act
@@ -64,5 +70,56 @@ describe('createClient', () => {
             // Assert
             expect(mergedOptions).toStrictEqual(expectedResult);
         });
+    });
+});
+
+describe('Http Interceptor', () => {
+    // Arrange
+    const expected = { test: 'jazz' };
+    console.log = jest.fn();
+    const mock = new MockAdapter(axios);
+    mock.onGet('/test').reply(200, expected);
+
+    it('should output api timing details to console if in development mode', async () => {
+        // Arrange
+        const config = {
+            baseUrl: 'https://www.mockendpoint.com',
+            timeout: 1000,
+            contentType: 'application/json',
+            isDevelopment: true
+        };
+        const httpClient = createClient(config);
+
+        // Act
+        const actual = await httpClient.get('/test');
+
+        // Assert
+        expect(actual.statusCode).toBe(200);
+        expect(actual.data).toEqual(expected);
+        expect(console.log).toHaveBeenCalledWith(expect.stringContaining('GET|/test|200|'));
+
+        console.log.mockClear();
+    });
+
+    it('should not output api timing details to console if not in development mode', async () => {
+        // Arrange
+        const config = {
+            baseUrl: 'https://www.mockendpoint.com',
+            timeout: 1000,
+            contentType: 'application/json',
+            isDevelopment: false
+        };
+        const httpClient = createClient(config);
+
+        // Act
+        const actual = await httpClient.get('/test');
+
+        // Assert
+        expect(actual.statusCode).toBe(200);
+        expect(actual.data).toEqual(expected);
+        expect(console.log).not.toBeCalled();
+
+        console.log.mockClear();
+        mock.restore();
     });
 });