+++ packages/components/organisms/f-checkout/CHANGELOG.md
@@ -3,6 +3,29 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+
+v0.115.0
+------------------------------
+*May 18, 2021*
+
+### Removed
+- `UserNote.vue` as now it is just another form field from `f-form-field` component
+
+### Updated
+- `f-form-field` to version 1.12.1 to include textarea input-type
+
+
+v0.114.0
+------------------------------
+*May 17, 2021*
+
+### Fixed
+- Error modal spacing so that the title and the close button don't overlap
+
+### Changed
+- Updated `f-mega-modal` version
+
+
 v0.113.0
 ------------------------------
 *May 12, 2021*