+++ packages/storybook/CHANGELOG.md
@@ -3,6 +3,13 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+v0.34.0
+------------------------------
+*May 14, 2021*
+
+### Changed
+- Bumped f-http to v0.4.1 from v0.4.0
+
 
 v0.33.0
 ------------------------------